+++ staging/src/k8s.io/apiextensions-apiserver/pkg/apiserver/schema/cel/library/urls.go
@@ -22,7 +22,7 @@ import (
 	"github.com/google/cel-go/cel"
 	"github.com/google/cel-go/common/types"
 	"github.com/google/cel-go/common/types/ref"
-	"k8s.io/apiextensions-apiserver/third_party/forked/celopenapi/model"
+	"k8s.io/apiextensions-apiserver/pkg/apiserver/schema/cel/model"
 )
 
 // URLs provides a CEL function library extension of URL parsing functions.
