+++ staging/src/k8s.io/apiextensions-apiserver/pkg/apiserver/schema/cel/model/value_test.go
@@ -1,16 +1,18 @@
-// Copyright 2020 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+Copyright 2022 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
 
 package model
 
@@ -322,9 +324,9 @@ func TestMapValueIsSet(t *testing.T) {
 
 func TestObjectValueEqual(t *testing.T) {
 	objType := NewObjectType("Notice", map[string]*DeclField{
-		"name":     &DeclField{Name: "name", Type: StringType},
-		"priority": &DeclField{Name: "priority", Type: IntType},
-		"message":  &DeclField{Name: "message", Type: StringType, defaultValue: "<eom>"},
+		"name":     {Name: "name", Type: StringType},
+		"priority": {Name: "priority", Type: IntType},
+		"message":  {Name: "message", Type: StringType, defaultValue: "<eom>"},
 	})
 	name := NewField(1, "name")
 	name.Ref = testValue(t, 2, "alert")
